+++ packages/mint-components/src/components/sqm-big-stat/sqm-big-stat.tsx
@@ -21,7 +21,7 @@ export class BigStat {
    *
    * @uiWidget StatTypeSelectWidget
    * @uiName Stat Type
-   * @uiOptions {"version": 1.1}
+   * @uiWidgetOptions {"version": 1.1}
    */
   @Prop() statType: string;