+++ packages/mint-components/src/components/sqm-leaderboard-rank/sqm-leaderboard-rank.tsx
@@ -47,7 +47,7 @@ export class LeaderboardRank {
   /**
    * @uiName Leaderboard Time Interval
    * @uiWidget DateRange
-   * @uiOptions {"allowPastDates":true, "months": 1}
+   * @uiWidgetOptions {"allowPastDates":true, "months": 1}
    */
   @Prop() interval: string;